ns: rewrite IDocumentInterface
